+++ packages/cli/src/cli.ts
@@ -1,10 +1,13 @@
 #!/usr/bin/env node
-import { fork } from 'child_process'
 import { resolve } from 'path'
+import { fork } from 'child_process'
 import * as yargs from 'yargs'
 import { Argv } from 'ssr-types'
+import { copyViteConfig, checkVite } from 'ssr-server-utils'
 
 const spinnerProcess = fork(resolve(__dirname, './spinner')) // 单独创建子进程跑 spinner 否则会被后续的 require 占用进程导致 loading 暂停
+const debug = require('debug')('ssr:cli')
+const start = Date.now()
 
 yargs
   .command('start', 'Start Server', {}, async (argv: Argv) => {
@@ -12,21 +15,40 @@ yargs
       message: 'start'
     })
     process.env.NODE_ENV = 'development'
-    const { parseFeRoutes, loadPlugin } = require('ssr-server-utils')
+    // 只有本地开发环境才会使用 Vite
+    process.env.BUILD_TOOL = argv.vite ? 'vite' : 'webpack'
+    if (argv.test) {
+      // 开发同学本地 link 测试用
+      process.env.TEST = '1'
+    }
+    if (process.env.BUILD_TOOL === 'vite') {
+      const result = await checkVite()
+      if (!result) {
+        process.exit(0)
+      }
+      await copyViteConfig()
+    }
+
+    const { parseFeRoutes, loadPlugin } = await import('ssr-server-utils')
+    debug(`require ssr-server-utils time: ${Date.now() - start} ms`)
     const plugin = loadPlugin()
+    debug(`loadPlugin time: ${Date.now() - start} ms`)
     await parseFeRoutes()
     spinnerProcess.send({
       message: 'stop'
     })
+    debug(`parseFeRoutes ending time: ${Date.now() - start} ms`)
     await plugin.clientPlugin?.start?.(argv)
+    debug(`clientPlugin ending time: ${Date.now() - start} ms`)
     await plugin.serverPlugin?.start?.(argv)
+    debug(`serverPlugin ending time: ${Date.now() - start} ms`)
   })
   .command('build', 'Build server and client files', {}, async (argv: Argv) => {
     spinnerProcess.send({
       message: 'start'
     })
     process.env.NODE_ENV = 'production'
-    const { parseFeRoutes, loadPlugin } = require('ssr-server-utils')
+    const { parseFeRoutes, loadPlugin } = await import('ssr-server-utils')
     const plugin = loadPlugin()
     await parseFeRoutes()
     spinnerProcess.send({
@@ -36,10 +58,11 @@ yargs
     await plugin.serverPlugin?.build?.(argv)
   })
   .command('deploy', 'Deploy function to aliyun cloud or tencent cloud', {}, async (argv: Argv) => {
-    const { loadPlugin } = require('ssr-server-utils')
+    const { loadPlugin } = await import('ssr-server-utils')
+
     const plugin = loadPlugin()
-    if (!plugin.serverPlugin.deploy) {
-      console.log('当前插件不支持 deploy 功能，请使用 ssr-plugin-faas 插件 并创建对应 yml 文件 参考 https://www.yuque.com/midwayjs/faas/migrate_egg 或扫码进群了解')
+    if (!plugin?.serverPlugin?.deploy) {
+      console.log('当前插件不支持 deploy 功能，请使用 ssr-plugin-midway 插件 参考 https://www.yuque.com/midwayjs/faas/migrate_egg 或扫码进群了解')
       return
     }
     process.env.NODE_ENV = 'production'