+++ packages/plugin-vue3/src/entry/store.ts
@@ -1,0 +1,12 @@
+import { createStore as createVuexStore } from 'vuex'
+
+// @ts-expect-error
+import * as store from '@/store/index.ts'
+
+function createStore () {
+  return createVuexStore(store)
+}
+
+export {
+  createStore
+}