+++ packages/server-utils/src/cwd.ts
@@ -1,59 +1,39 @@
 import * as fs from 'fs'
 import { promises } from 'fs'
-import { join, isAbsolute } from 'path'
-import { UserConfig } from 'ssr-types'
+import { resolve } from 'path'
+import { UserConfig, IPlugin } from 'ssr-types'
 
 const getCwd = () => {
-  const cwd = process.cwd()
-  if (process.env.APP_ROOT) {
-    // avoid repeat cwd path
-    if (!isAbsolute(process.env.APP_ROOT)) {
-      return join(cwd, process.env.APP_ROOT)
-    }
-    return process.env.APP_ROOT
-  }
-  return cwd
+  return resolve(process.cwd(), process.env.APP_ROOT ?? '')
 }
 
 const getFeDir = () => {
-  // fe component folder path
-  const cwd = getCwd()
-  if (process.env.FE_ROOT) {
-    // avoid repeat cwd path
-    if (!isAbsolute(process.env.FE_ROOT)) {
-      return join(cwd, process.env.FE_ROOT)
-    }
-    return process.env.FE_ROOT
-  }
-  return join(cwd, './web')
+  return resolve(getCwd(), process.env.FE_ROOT ?? 'web')
 }
 
 const getPagesDir = () => {
-  return join(getFeDir(), './pages')
-}
-
-const getVuexStoreFilePath = () => {
-  return join(getFeDir(), './store')
+  return resolve(getFeDir(), 'pages')
 }
 
 const getUserConfig = (): UserConfig => {
   // 生产环境如果有 config.prod 则读取
   const isProd = process.env.NODE_ENV === 'production'
-  const hasProdConfig = fs.existsSync(join(getCwd(), './config.prod.js'))
-  return require(join(getCwd(), isProd && hasProdConfig ? './config.prod' : './config'))
+  const hasProdConfig = fs.existsSync(resolve(getCwd(), 'config.prod.js'))
+  return require(resolve(getCwd(), isProd && hasProdConfig ? 'config.prod' : 'config'))
 }
 
-const loadPlugin = () => {
-  return require(join(getCwd(), './plugin'))
+const loadPlugin = (): IPlugin => {
+  return require(resolve(getCwd(), 'plugin'))
 }
+
 const isFaaS = async () => {
-  const result = await promises.access(join(getCwd(), './f.yml'))
+  const result = await promises.access(resolve(getCwd(), 'f.yml'))
     .then(() => true)
     .catch(() => false)
   return result
 }
 
-const getLocalNodeModules = () => join(__dirname, '../../../')
+const getLocalNodeModules = () => resolve(__dirname, '../../../')
 
 const processError = (err: any) => {
   if (err) {
@@ -61,6 +41,32 @@ const processError = (err: any) => {
     process.exit(1)
   }
 }
+const accessFile = async (file: string) => {
+  const result = await promises.access(file)
+    .then(() => true)
+    .catch(() => false)
+  return result
+}
+
+const checkVite = async () => {
+  const result = await accessFile(resolve(getCwd(), './node_modules/vite/package.json'))
+  if (!result) {
+    console.log('当前项目缺少 vite 依赖，请根据实际技术栈安装 vite @vitejs/plugin-vue 或 其他对应插件')
+    return false
+  }
+  return true
+}
+
+const copyViteConfig = async () => {
+  // 如果当前项目没有 vite.config 则复制默认的文件
+  const result = await accessFile(resolve(getCwd(), './vite.config.js'))
+  if (!result) {
+    console.log('检测到当前目录缺少 vite.config.js 文件，自动创建默认模版')
+
+    await promises.copyFile(resolve(getCwd(), './node_modules/ssr-plugin-vue3/src/config/vite.config.tpl'), resolve(getCwd(), './vite.config.js'))
+  }
+}
+
 export {
   getCwd,
   getFeDir,
@@ -70,5 +76,7 @@ export {
   loadPlugin,
   getLocalNodeModules,
   processError,
-  getVuexStoreFilePath
+  accessFile,
+  copyViteConfig,
+  checkVite
 }